+++ zetaclient/chains/evm/signer/outbound_data.go
@@ -29,9 +29,8 @@ type OutboundData struct {
 	asset  ethcommon.Address
 	amount *big.Int
 
-	gas    Gas
-	nonce  uint64
-	height uint64
+	gas   Gas
+	nonce uint64
 
 	message []byte
 
@@ -53,7 +52,6 @@ type OutboundData struct {
 func NewOutboundData(
 	ctx context.Context,
 	cctx *types.CrossChainTx,
-	height uint64,
 	logger zerolog.Logger,
 ) (*OutboundData, bool, error) {
 	if cctx == nil {
@@ -139,9 +137,8 @@ func NewOutboundData(
 		asset:  ethcommon.HexToAddress(cctx.InboundParams.Asset),
 		amount: outboundParams.Amount.BigInt(),
 
-		gas:    gas,
-		nonce:  outboundParams.TssNonce,
-		height: height,
+		gas:   gas,
+		nonce: outboundParams.TssNonce,
 
 		message: message,