+++ zetaclient/testutils/mocks/zetacore.go
@@ -12,6 +12,8 @@ import (
 
 	context "context"
 
+	cosmos_sdktypes "github.com/cosmos/cosmos-sdk/types"
+
 	crosschaintypes "github.com/zeta-chain/node/x/crosschain/types"
 
 	errors "github.com/zeta-chain/node/pkg/errors"
@@ -168,9 +170,9 @@ func (_m *ZetacoreClient) GetBaseGasPrice(ctx context.Context) (int64, error) {
 	return r0, r1
 }
 
-// GetBlockHeight provides a mock function with given fields: ctx
-func (_m *ZetacoreClient) GetBlockHeight(ctx context.Context) (int64, error) {
-	ret := _m.Called(ctx)
+// GetBlockHeight provides a mock function with given fields: _a0
+func (_m *ZetacoreClient) GetBlockHeight(_a0 context.Context) (int64, error) {
+	ret := _m.Called(_a0)
 
 	if len(ret) == 0 {
 		panic("no return value specified for GetBlockHeight")
@@ -179,16 +181,16 @@ func (_m *ZetacoreClient) GetBlockHeight(ctx context.Context) (int64, error) {
 	var r0 int64
 	var r1 error
 	if rf, ok := ret.Get(0).(func(context.Context) (int64, error)); ok {
-		return rf(ctx)
+		return rf(_a0)
 	}
 	if rf, ok := ret.Get(0).(func(context.Context) int64); ok {
-		r0 = rf(ctx)
+		r0 = rf(_a0)
 	} else {
 		r0 = ret.Get(0).(int64)
 	}
 
 	if rf, ok := ret.Get(1).(func(context.Context) error); ok {
-		r1 = rf(ctx)
+		r1 = rf(_a0)
 	} else {
 		r1 = ret.Error(1)
 	}
@@ -1008,6 +1010,36 @@ func (_m *ZetacoreClient) PostVoteTSS(_a0 context.Context, tssPubKey string, key
 
 	if rf, ok := ret.Get(1).(func(context.Context, string, int64, chains.ReceiveStatus) error); ok {
 		r1 = rf(_a0, tssPubKey, keyGenZetaHeight, _a3)
+	} else {
+		r1 = ret.Error(1)
+	}
+
+	return r0, r1
+}
+
+// QueryTxResult provides a mock function with given fields: hash
+func (_m *ZetacoreClient) QueryTxResult(hash string) (*cosmos_sdktypes.TxResponse, error) {
+	ret := _m.Called(hash)
+
+	if len(ret) == 0 {
+		panic("no return value specified for QueryTxResult")
+	}
+
+	var r0 *cosmos_sdktypes.TxResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(string) (*cosmos_sdktypes.TxResponse, error)); ok {
+		return rf(hash)
+	}
+	if rf, ok := ret.Get(0).(func(string) *cosmos_sdktypes.TxResponse); ok {
+		r0 = rf(hash)
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(*cosmos_sdktypes.TxResponse)
+		}
+	}
+
+	if rf, ok := ret.Get(1).(func(string) error); ok {
+		r1 = rf(hash)
 	} else {
 		r1 = ret.Error(1)
 	}